+++ packages/core/src/common/CommonTypes.ts
@@ -1,2 +1,9 @@
 /**  Unknown object */
 export type UnknownObject = Record<string, unknown>;
+
+export type JSONValue =
+  | string
+  | number
+  | boolean
+  | { [key: string]: JSONValue }
+  | Array<JSONValue>;