+++ packages/core/src/variant-handlers/VariantHandlers.js
@@ -8,11 +8,11 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const Json = require("./handlers/Json");
-const Text = require("./handlers/Text");
-const Middleware = require("./handlers/Middleware");
-const Static = require("./handlers/Static");
-const Status = require("./handlers/Status");
+const { VariantHandlerJson } = require("./handlers/Json");
+const { VariantHandlerText } = require("./handlers/Text");
+const { VariantHandlerMiddleware } = require("./handlers/Middleware");
+const { VariantHandlerStatic } = require("./handlers/Static");
+const { VariantHandlerStatus } = require("./handlers/Status");
 const { VariantHandlerFile } = require("./handlers/File");
 
 const OPTIONS = [
@@ -32,7 +32,14 @@ class VariantHandlers {
   constructor({ logger, config }) {
     this._logger = logger;
     this._registeredVariantHandlers = [];
-    this._coreVariantHandlers = [Json, Text, Status, Middleware, Static, VariantHandlerFile];
+    this._coreVariantHandlers = [
+      VariantHandlerJson,
+      VariantHandlerText,
+      VariantHandlerStatus,
+      VariantHandlerMiddleware,
+      VariantHandlerStatic,
+      VariantHandlerFile,
+    ];
     this._config = config;
 
     [this._registerOption] = this._config.addOptions(OPTIONS);