+++ packages/core/src/variant-handlers/VariantHandlersTypes.ts
@@ -8,9 +8,9 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type { UnknownObject } from "../common/CommonTypes";
+import type { UnknownObject, JSONValue } from "../common/CommonTypes";
 import type { JSONSchema7WithInstanceof } from "../mock/ValidationsTypes";
-import type { NextFunction, Request, Response } from "../server/ServerTypes";
+import type { NextFunction, Request, Response, RequestHandler } from "../server/ServerTypes";
 
 export type VariantHandlerBaseConstructorOptions = UnknownObject;
 
@@ -21,7 +21,7 @@ export interface VariantHandlerResponsePreview {
   /** Response headers */
   headers?: UnknownObject;
   /** Response body */
-  body?: UnknownObject | string | null;
+  body?: string | JSONValue | null;
 }
 
 /** Common interface of variant handler constructors */
@@ -42,7 +42,7 @@ export interface VariantHandlerBaseInterface {
   get preview(): VariantHandlerResponsePreview | null;
 }
 
-/** Common interface of variant handlers. Variant handlers should be created extending this interface */
+/** Common interface of variant handlers of type middleware. Variant handlers providing a middleware should be created extending this interface */
 export interface VariantHandlerBaseInterfaceWithMiddleware extends VariantHandlerBaseInterface {
   /**
    * Express middleware to be executed when the request is received for a route using this variant handler
@@ -51,4 +51,12 @@ export interface VariantHandlerBaseInterfaceWithMiddleware extends VariantHandle
    * @param next - Express method for executing next middleware {@link NextFunction}
    */
   middleware(req: Request, res: Response, next: NextFunction): void;
+}
+
+/** Common interface of variant handlers of type router. Variant handlers providing a router should be created extending this interface */
+export interface VariantHandlerBaseInterfaceWithRouter extends VariantHandlerBaseInterface {
+  /**
+   * Returns an Express router to be mounted in the route path
+   */
+  get router(): RequestHandler;
 }