+++ packages/core/src/variant-handlers/handlers/Static.ts
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 Javier Brea
+Copyright 2023 Javier Brea
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -8,16 +8,33 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-"use strict";
+import type { LoggerInterface } from "@mocks-server/logger";
+import { static as expressStatic } from "express";
+import type { ServeStaticOptions } from "serve-static";
 
-const express = require("express");
+import type {
+  VariantHandlerStaticConstructor,
+  VariantHandlerStaticInterface,
+  VariantHandlerStaticOptions,
+  VariantHandlerStaticPreview,
+} from "./StaticTypes";
 
-class Static {
-  static get id() {
+import type { CoreInterface } from "../../CoreTypes";
+import type { JSONSchema7WithInstanceof } from "../../mock/ValidationsTypes";
+import type { RequestHandler, Response } from "../../server/ServerTypes";
+
+export const VariantHandlerStatic: VariantHandlerStaticConstructor = class VariantHandlerStatic
+  implements VariantHandlerStaticInterface
+{
+  private _options: VariantHandlerStaticOptions;
+  private _logger: LoggerInterface;
+  private _expressStaticOptions: ServeStaticOptions;
+
+  static get id(): string {
     return "static";
   }
 
-  static get validationSchema() {
+  static get validationSchema(): JSONSchema7WithInstanceof {
     return {
       type: "object",
       properties: {
@@ -36,27 +53,29 @@ class Static {
     };
   }
 
-  constructor(options, core) {
+  constructor(options: VariantHandlerStaticOptions, core: CoreInterface) {
     this._options = options;
     this._expressStaticOptions = this._options.options || {};
     this._logger = core.logger;
-    this._core = core;
   }
 
-  get router() {
+  get router(): RequestHandler {
     this._logger.verbose(`Creating router to serve static folder ${this._options.path}`);
 
     let setHeadersOption;
     if (this._options.headers) {
-      setHeadersOption = (res) => {
+      setHeadersOption = (res: Response): void => {
         res.set(this._options.headers);
       };
     }
-    return express.static(this._options.path, {
+
+    return expressStatic(this._options.path, {
       ...this._expressStaticOptions,
       setHeaders: this._expressStaticOptions.setHeaders || setHeadersOption,
     });
   }
-}
 
-module.exports = Static;
+  get preview(): VariantHandlerStaticPreview {
+    return null;
+  }
+};