+++ packages/core/src/variant-handlers/handlers/Text.ts
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 Javier Brea
+Copyright 2023 Javier Brea
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -8,14 +8,37 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-"use strict";
+import type { LoggerInterface } from "@mocks-server/logger";
 
-class Json {
+import type {
+  VariantHandlerTextConstructor,
+  VariantHandlerTextInterface,
+  VariantHandlerTextOptions,
+  VariantHandlerTextPreview,
+} from "./TextTypes";
+
+import type { UnknownObject } from "../../common/CommonTypes";
+import type { CoreInterface } from "../../CoreTypes";
+import type { JSONSchema7WithInstanceof } from "../../mock/ValidationsTypes";
+import type { Request, Response } from "../../server/ServerTypes";
+
+export const VariantHandlerText: VariantHandlerTextConstructor = class VariantHandlerText
+  implements VariantHandlerTextInterface
+{
+  private _options: VariantHandlerTextOptions;
+  private _logger: LoggerInterface;
+
   static get id() {
-    return "json";
+    return "text";
   }
 
-  static get validationSchema() {
+  get defaultHeaders(): UnknownObject {
+    return {
+      "Content-Type": "text/plain; charset=utf-8",
+    };
+  }
+
+  static get validationSchema(): JSONSchema7WithInstanceof {
     return {
       type: "object",
       properties: {
@@ -26,14 +49,7 @@ class Json {
           type: "number",
         },
         body: {
-          oneOf: [
-            {
-              type: "object",
-            },
-            {
-              type: "array",
-            },
-          ],
+          type: "string",
         },
       },
       required: ["status", "body"],
@@ -41,19 +57,12 @@ class Json {
     };
   }
 
-  get defaultHeaders() {
-    return {
-      "Content-Type": "application/json; charset=utf-8",
-    };
-  }
-
-  constructor(options, core) {
+  constructor(options: VariantHandlerTextOptions, core: CoreInterface) {
     this._options = options;
     this._logger = core.logger;
-    this._core = core;
   }
 
-  middleware(req, res) {
+  middleware(req: Request, res: Response): void {
     this._logger.debug(`Setting headers | req: ${req.id}`);
     res.set({ ...this.defaultHeaders, ...this._options.headers });
     res.status(this._options.status);
@@ -61,12 +70,10 @@ class Json {
     res.send(this._options.body);
   }
 
-  get preview() {
+  get preview(): VariantHandlerTextPreview {
     return {
       body: this._options.body,
       status: this._options.status,
     };
   }
-}
-
-module.exports = Json;
+};