+++ packages/core/test/components/mock/Mock.spec.js
@@ -19,7 +19,7 @@ const { Alerts } = require("../../../src/alerts/Alerts");
 const Mock = require("../../../src/mock/Mock");
 const ConfigMock = require("../common/Config.mocks");
 
-const JsonRoutesHandler = require("../../../src/variant-handlers/handlers/Json");
+const { VariantHandlerJson } = require("../../../src/variant-handlers/handlers/Json");
 
 describe("Mock", () => {
   let configMock;
@@ -57,7 +57,7 @@ describe("Mock", () => {
     mock = new Mock(methods, core);
     mock._routesLoaders.contents = [];
     mock._collectionsLoaders.contents = [];
-    mock.init([JsonRoutesHandler]);
+    mock.init([VariantHandlerJson]);
   });
 
   afterEach(() => {